requirements.txt
lz4
paho-mqtt
